src/ac_training_lab/ot-2/requirements-mqtt.txt
# Sparkplug B MQTT orchestration requirements
# Compatible with Opentrons (no pydantic/anyio/jsonschema conflicts)

paho-mqtt>=1.6.1
mqtt-spb-wrapper>=1.0.0
